split off token class
